deploy: ae3b769049d2b4212bd3765dac464789ed6383c6

+++ OgreMain/src/GLX/OgreTimer.cpp
@@ -53,7 +53,7 @@ uint64 Timer::getMilliseconds()
 {
     struct timeval now;
     gettimeofday(&now, NULL);
-    return (now.tv_sec-start.tv_sec)*1000+(now.tv_usec-start.tv_usec)/1000;
+    return (now.tv_sec-start.tv_sec)*1000ul+(unsigned long)((now.tv_usec-start.tv_usec)/1000l);
 }
 
 //--------------------------------------------------------------------------------//
@@ -61,7 +61,7 @@ uint64 Timer::getMicroseconds()
 {
     struct timeval now;
     gettimeofday(&now, NULL);
-    return (now.tv_sec-start.tv_sec)*1000000+(now.tv_usec-start.tv_usec);
+    return (now.tv_sec-start.tv_sec)*1000000ul+(now.tv_usec-start.tv_usec);
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
